+++ src/pages/Shuttle/ConfirmModal/Approve/Out.js
@@ -1,0 +1,2 @@
+function Out() {}
+export default Out
