+++ src/pages/Shuttle/ConfirmModal/Approve/index.js
@@ -1,0 +1,2 @@
+export {default as ApproveIn} from './In'
+export {default as ApproveOut} from './Out'
